requirements.txt
watchdog
pandas
twilio
pyyaml
Flask
transformers
torch
requests